% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{logisticregSGD}
\alias{logisticregSGD}
\alias{LogiSGD}
\alias{LogiSGDCI}
\title{parameter estimation and confidence interval construction for logistic regression with SGD}
\usage{
LogiSGD(X, Y, lr0, C = 2/3, burnin = 0.2)

LogiSGDCI(
  X,
  Y,
  lr0,
  C = 2/3,
  burnin = 0.2,
  B = 200,
  per_dis = "Poi",
  CI.level = 0.95,
  CI.method = "sd"
)
}
\arguments{
\item{lr0}{lr0 and C are used to calcluate learning rate, learning rate for k-th data point is
calculated as lr_k = lr_0*k **(-C)}

\item{C}{see description for lr0 and C is between 0.5 and 1}

\item{burnin}{a number between 0 and 1; this represent the proportion of estimates
that are not used for calculating final estimate}

\item{B}{number of perturbed estimates generated}

\item{CI.level}{confidence level for constructing confidence interval}

\item{CI.method}{the method to construct confidence interval
it takes one of the following values 'sd' or 'quantile'
'sd' means using sample standard deviation of perturbed
estimates to estimate the standard deviation of SGD estimate
'quantile' means using the upper alpha/2 and lower alpha/2
quantile of the perturbed estimates to construct
confidence interval}

\item{x}{matrix for independent variable, does not include the constant column}

\item{y}{response variable, it takes values of 1 or -1}

\item{Per_dis}{the distribution to generate perturbation parameter
it takes the following values, 'Poi': Poission distribution, Poi(1)
'exp', exponential distribution, exp(1)}
}
\value{
\strong{\code{LogiSGD}} \cr
This function will return estimated coefficients for logistic regression. \cr
\strong{\code{LogiSGDCI}} \cr
This function will return estimated coefficients and confidence intervals for logistic regression. \cr
}
\description{
This for applying stochastic gradient descent (SGD) for parameter estimation of
logistic regression model. \code{LogiSGD} apply SGD for parameter estimation and \code{LogiSGDCI}
apply SGD for parameter estimation and also construct confidence intervals for estimates.
}
\details{
Parameter estimation and confidence interval construction are based on Example 2 of Yixin Fang, Jinfeng Xu, and Lei Yang.
Online bootstrap confidence intervals for the stochastic gradient descent estimator.
The Journal of Machine Learning Research, 19(1):3053-3073, 2018.
}
\examples{
\dontrun{
generate x and y
N<-10000
beta_0<-0.1
beta_1<-0.1
beta_2<--0.1
SD<-1
x1<-rnorm(N, 0, 1)
x2<-rnorm(N,0,1)
x<-matrix(c(x1,x2), nrow=N, byrow = FALSE)
EXP<-exp(beta_0+beta_1*x1+beta_2*x2)
P<- EXP/(1+EXP)
yGLM<-ifelse(runif(N)<P,1,0)
ySGD<-ifelse(yGLM==1, 1, -1)
LogiSGD(x,ySGD,lr0=0.1, C=2/3, burnin=0.2)
LogiSGDCI(x,ySGD,lr0=0.1, C=2/3, burnin=0.2, B=200, CI.method = 'sd', per_dis = 'exp')
glm function provided by r as a control
glm(yGLM ~ x1 + x2, family='binomial'(link='logit'))
}

}
\references{
Yixin Fang, Jinfeng Xu, and Lei Yang.
Online bootstrap confidence intervals for the stochastic gradient descent estimator.
The Journal of Machine Learning Research, 19(1):3053-3073, 2018.
}
