% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAR.R
\name{SAR}
\alias{SAR}
\alias{SARSGD}
\title{parameter estimation and confidence interval construction for spatial autoregressive (SAR) model with SGD}
\usage{
SARSGD(x, y, W, lr0, C = 2/3, burnin = 0.2)
}
\description{
This for applying stochastic gradient descent (SGD) for parameter estimation of
SAR model. \code{SGD} apply SGD for parameter estimation and \code{SGDCI}
}
